+++ AdapterUnitTests/PangleAdapterTests/AUTPangleAppOpenAdTests.m
@@ -47,6 +47,9 @@
   OCMVerifyAll(_request);
   OCMVerifyAll(_ad);
   GADMobileAds.sharedInstance.requestConfiguration.tagForChildDirectedTreatment = nil;
+  GADMobileAds.sharedInstance.requestConfiguration.tagForUnderAgeOfConsent = nil;
+  GADMobileAds.sharedInstance.requestConfiguration.ageRestrictedTreatment =
+      GADAgeRestrictedTreatmentUnspecified;
 }
 
 - (nonnull AUTKMediationAppOpenAdEventDelegate *)loadAdWithPlacementID:
@@ -120,6 +123,40 @@
 
 - (void)testLoadAdForNonChildAudience {
   GADMobileAds.sharedInstance.requestConfiguration.tagForChildDirectedTreatment = @NO;
+  [self loadAdWithPlacementID:@"ID"];
+}
+
+- (void)testLoadAdForChildAudienceWithTagForUnderAgeOfConsent {
+  GADMobileAds.sharedInstance.requestConfiguration.tagForUnderAgeOfConsent = @YES;
+  NSError *expectedError = [[NSError alloc] initWithDomain:GADMAdapterPangleErrorDomain
+                                                      code:GADPangleErrorChildUser
+                                                  userInfo:nil];
+  [self loadAdFailureWithPlacementID:@"ID" expectedError:expectedError];
+}
+
+- (void)testLoadAdForNonChildAudienceWithTagForUnderAgeOfConsent {
+  GADMobileAds.sharedInstance.requestConfiguration.tagForUnderAgeOfConsent = @NO;
+  [self loadAdWithPlacementID:@"ID"];
+}
+
+- (void)testLoadAdForChildAudienceWithAgeRestrictedTreatment {
+  GADMobileAds.sharedInstance.requestConfiguration.ageRestrictedTreatment =
+      GADAgeRestrictedTreatmentChild;
+  NSError *expectedError = [[NSError alloc] initWithDomain:GADMAdapterPangleErrorDomain
+                                                      code:GADPangleErrorChildUser
+                                                  userInfo:nil];
+  [self loadAdFailureWithPlacementID:@"ID" expectedError:expectedError];
+}
+
+- (void)testLoadAdForNonChildAudienceWithAgeRestrictedTreatment {
+  GADMobileAds.sharedInstance.requestConfiguration.ageRestrictedTreatment =
+      GADAgeRestrictedTreatmentTeen;
+  [self loadAdWithPlacementID:@"ID"];
+}
+
+- (void)testLoadAdForNonChildAudienceWithAgeRestrictedTreatmentUnspecified {
+  GADMobileAds.sharedInstance.requestConfiguration.ageRestrictedTreatment =
+      GADAgeRestrictedTreatmentUnspecified;
   [self loadAdWithPlacementID:@"ID"];
 }