+++ AdapterUnitTests/PangleAdapterTests/AUTPangleNativeAdTests.m
@@ -74,6 +74,9 @@
   OCMVerifyAll(_data);
   OCMVerifyAll(_icon);
   GADMobileAds.sharedInstance.requestConfiguration.tagForChildDirectedTreatment = nil;
+  GADMobileAds.sharedInstance.requestConfiguration.tagForUnderAgeOfConsent = nil;
+  GADMobileAds.sharedInstance.requestConfiguration.ageRestrictedTreatment =
+      GADAgeRestrictedTreatmentUnspecified;
 }
 
 - (nonnull AUTKMediationNativeAdEventDelegate *)loadAdWithPlacementID:
@@ -147,6 +150,40 @@
 
 - (void)testLoadAdForNonChildAudience {
   GADMobileAds.sharedInstance.requestConfiguration.tagForChildDirectedTreatment = @NO;
+  [self loadAdWithPlacementID:@"12345"];
+}
+
+- (void)testLoadAdForUnderAgeOfConsent {
+  GADMobileAds.sharedInstance.requestConfiguration.tagForUnderAgeOfConsent = @YES;
+  NSError *expectedError = [[NSError alloc] initWithDomain:GADMAdapterPangleErrorDomain
+                                                      code:GADPangleErrorChildUser
+                                                  userInfo:nil];
+  [self loadAdFailureWithPlacementID:@"ID" expectedError:expectedError];
+}
+
+- (void)testLoadAdForNonUnderAgeOfConsent {
+  GADMobileAds.sharedInstance.requestConfiguration.tagForUnderAgeOfConsent = @NO;
+  [self loadAdWithPlacementID:@"12345"];
+}
+
+- (void)testLoadAdForAgeRestrictedChild {
+  GADMobileAds.sharedInstance.requestConfiguration.ageRestrictedTreatment =
+      GADAgeRestrictedTreatmentChild;
+  NSError *expectedError = [[NSError alloc] initWithDomain:GADMAdapterPangleErrorDomain
+                                                      code:GADPangleErrorChildUser
+                                                  userInfo:nil];
+  [self loadAdFailureWithPlacementID:@"ID" expectedError:expectedError];
+}
+
+- (void)testLoadAdForAgeRestrictedTeen {
+  GADMobileAds.sharedInstance.requestConfiguration.ageRestrictedTreatment =
+      GADAgeRestrictedTreatmentTeen;
+  [self loadAdWithPlacementID:@"12345"];
+}
+
+- (void)testLoadAdForAgeRestrictedUnspecified {
+  GADMobileAds.sharedInstance.requestConfiguration.ageRestrictedTreatment =
+      GADAgeRestrictedTreatmentUnspecified;
   [self loadAdWithPlacementID:@"12345"];
 }