+++ adapters/Line/CHANGELOG.md
@@ -1,5 +1,15 @@
 ## LINE iOS Mediation Adapter Changelog
 
+#### Next Version
+- Resize the Line adapter's information icon image.
+
+#### [Version 3.0.1.1](https://dl.google.com/googleadmobadssdk/mediation/ios/line/LineAdapter-3.0.1.1.zip)
+- Adapter now forwards the `ageRestrictedTreatment` Google Mobile Ads SDK flag to FiveAd SDK's COPPA API.
+
+Built and tested with:
+- Google Mobile Ads SDK version 13.3.0.
+- FiveAd SDK version 3.0.1.
+
 #### [Version 3.0.1.0](https://dl.google.com/googleadmobadssdk/mediation/ios/line/LineAdapter-3.0.1.0.zip)
 - Verified compatibility with FiveAd SDK version 3.0.1.